Fix pdot assignment in populate.py
In mohak300501 fork a plot of P-Pdot diagram was added at the end. But populate.py didn't assign Pdot so a line was added to give pdot values. This resulted in a transcription error where instead of values like (10^12)^2 it was determining 12^2, meaning Pdot values were 1e-37 rather than 1e-15.

This doesn't effect dosurvey as luminosity is independent of Pdot (and P and everything else). This might cause a problem if running evolve.py but maybe not as it seems to calculate pdot again based on period and this same relation which assumes n=3.

Indicative plot now looks better.